fix for elapsed time: changed data type to double, round to 2 digits

+++ net.bhl.cdt.model.cabin/src/net/bhl/cdt/model/astar/StopWatch.java
@@ -61,8 +61,8 @@ public class StopWatch {
 	 * 
 	 * @return the time in milliseconds
 	 */
-	public int getElapsedTimeSecs() {
-		int elapsed;
+	public long getElapsedTimeSecs() {
+		long elapsed;
 		if (running) {
 			elapsed = (int) ((System.currentTimeMillis() - startTime) / 1000);
 		} else {
